+++ tests/pickle_test.py
@@ -185,7 +185,7 @@ class PickleTest(jtu.JaxTestCase):
   def test_pickle_op_sharding_sharding(self):
     op_sharding = xla.xc.OpSharding()
     op_sharding.type = xla.xc.OpSharding.Type.REPLICATED
-    s = jax.sharding.OpShardingSharding(jax.devices(), op_sharding)
+    s = jax.sharding.GSPMDSharding(jax.devices(), op_sharding)
     self.assertEqual(s, pickle.loads(pickle.dumps(s)))
 
   @unittest.skipIf(cloudpickle is None, "Requires cloudpickle")